Move spin1 interpreter-specific code into new file
